Update README.md
Added a note for a very common use case; using Session in conjunction with PassportJS. In this configuration, Passport always adds an object to the uninitialized session object, causing it to be saved, even if saveUninitialized has been set. It's a very common pitfall, and I spent a while trying to work out why saveUninitialized was not having the effect as described.