feat: Option to allow regenerate() to preserve old session
Add preserve option to session.regenerate() function for use cases
when it is desirable to regenerate the session, but not destroy
the old session immediately. For example when there could be
multiple ajax requests inflight with the old session id.

Existing functionality is preserved if the option is omitted.
Existing test was improved to cover session destuction and new
test was added for the preserve case.