Make variable type checks consistent
Whenever you use the `typeof` operator, you are guaranteed to be working with
strings, as this is what typeof returns. For this reason it is generally a good
idea to use the equality operator instead of the strict equality operator for
comparisons.

This is consistent with other areas of the code where the equality operator is
already in use.