Added a passReqToStore option, used instead of function arguments length to determine if the request should be passed to the store or not.
To that end, also exposed the session options in the request, and moved the load method so that it is overridden in the request, instead of defined at the store, i.e. all stores get it "for free".

Tweaked the "should set maxAge" cookie test to store the time at the start, making it non flaky.

Added more tests.